renaming and updating imports for TurnInfo, TurnView and TurnScore
