+++ src/components/TurnScore/index.js
@@ -1,0 +1,7 @@
+import React from "react"
+
+function TurnScore() {
+  return <div>5 aciertos</div>
+}
+
+export default TurnScore
