--- src/components/turnInfo/TurnInfo.spec.js
@@ -1,1 +1,0 @@
-test("", () => {})
