--- src/components/turnScore/TurnScore.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function TurnScore() {
-	return (
-		<div>
-            5 aciertos
-		</div>
-	)
-}
-
-export default TurnScore
