--- src/components/turnScore/TurnScore.spec.js
@@ -1,1 +1,0 @@
-test("", () => {})
