--- src/components/turnView/TurnView.spec.js
@@ -1,1 +1,0 @@
-test("", () => {})
